+++ test/magma/document_types/prompt_test.exs
@@ -25,18 +25,15 @@ defmodule Magma.PromptTest do
 
              # #{prompt.name}
 
-             ## System prompt
+             ## Context
 
-             ![[Magma.system.config#Persona|]]
+             ### Knowledge Base
 
-             ### Context knowledge
+             Read the following documents carefully:
 
-             The following sections contain background knowledge you need to be aware of.
+             -
 
-             ![[Magma.system.config#Context knowledge|]]
-
-
-             ## Request
+             ## Task
 
              """