+++ test/magma/generation/manual_test.exs
@@ -63,11 +63,11 @@ defmodule Magma.Generation.ManualTest do
              """
              # #{prompt.name}
 
-             ## #{Prompt.Template.system_prompt_section_title()}
+             ## #{Prompt.Template.context_section_title()}
 
              #{system_prompt}
 
-             ## #{Prompt.Template.request_prompt_section_title()}
+             ## #{Prompt.Template.task_section_title()}
 
              #{request_prompt}
              """